A Whole Buncha Stuff
Switched out PSE for Income because PSE was not percent change and I could not find the annual data for % change PSE. Income makes sense though.

Did a complete analysis of realtionship with genres, started looking at GDP (made scatterplot), but realized I wanted to circle back to see the impact of # of movies made and # tickets sold.